travis: add ARM64 support

+++ tools/travis-ci/run.sh
@@ -80,16 +80,17 @@ if [ "${TRAVIS_OS_NAME}" == "osx" ]; then
 	GROK_CXX_VERSION=xcode${GROK_CXX_VERSION:6}
 elif [ "${TRAVIS_OS_NAME}" == "linux" ]; then
 	GROK_OS_NAME=linux
-	
-	# We need a newer version of cmake than travis-ci provides
-	wget --no-check-certificate -qO - https://cmake.org/files/v3.17/cmake-3.17.0-Linux-x86_64.tar.gz | tar -xz
-	# copy to a directory that will not changed every version
-	mv cmake-3.17.0-Linux-x86_64 cmake-install
-		
 	if which lsb_release > /dev/null; then
 		GROK_OS_NAME=$(lsb_release -si)$(lsb_release -sr | sed 's/\([^0-9]*\.[0-9]*\).*/\1/')
 	fi
-	sudo unlink /usr/bin/g++ && sudo ln -s /usr/bin/g++-10 /usr/bin/g++
+
+	if  [ "${TRAVIS_CPU_ARCH}" == "amd64" ]; then	
+		# We need a newer version of cmake than travis-ci provides
+		wget --no-check-certificate -qO - https://cmake.org/files/v3.17/cmake-3.17.0-Linux-x86_64.tar.gz | tar -xz
+		# copy to a directory that will not changed every version
+		mv cmake-3.17.0-Linux-x86_64 cmake-install
+		sudo unlink /usr/bin/g++ && sudo ln -s /usr/bin/g++-10 /usr/bin/g++
+	fi
 	g++ --version
 	if [ -z "${CXX##*g++*}" ]; then
 		GROK_CXX_VERSION=$(${CXX} --version | head -1 | sed 's/.*\ \([0-9.]*[0-9]\)/\1/')
